requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0
#rust based linter
ruff==0.0.284
# add necessary databricks req
pandas
python-dotenv
databricks-sql-connector
random